pyproject.toml
[project]
name = "eval_rag"
version = "0.1.0"
description = "Audit LLM"
license = "BSD-3-Clause"
requires-python = ">=3.11,<3.13"
authors = [
    { name = "Grégoire Martinon", email = "gregoire.martinon@capgemini.com" },
    { name = "Jérôme Bohard", email = "jerome.bohard@capgemini.com" },
    { name = "Antoine Lojou", email = "antoine.lojou@capgemini.com" },
    { name = "Alexandra Lorenzo", email = "alexandra.lorenzo-de-brionne@capgemini.com" },
]
readme = "README.md"
dependencies = [
    "langchain==0.3.17",
    "numpy==1.26.4",
    "openai==1.58.1",
    "pandas==2.2.2",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "ppi-python==0.2.3",
    "langfuse==2.59.3",
    "pydantic==2.10.4",
    "scipy>=1.15.2",
]

[dependency-groups]
dev = ["pytest==8.3.3", "ruff==0.6.7", "mypy==1.15.0", "pre-commit==4.2.0"]
explo = [
    "plotly==6.0.1",
    "notebook==7.3.2",
]
